+++ CHANGELOG.md
@@ -665,3 +665,11 @@
 - Add experimental ActivateTaggedEntitesSpecial(Ex)
 - Add ScaleBy function to VectorUtils
 - Add GameParams.DialogueVnSpeakerCaseSensitive option
+
+# Rnext
+
+- Fix UpdatePlayerModelOnConfigChange throwing an exception if game state doesn't exist
+- Add AlwaysSetXScale flag to FacingSprite components and SetSpriteOnQuad (works around issues with sprite not changing when object is instantaneously flipped)
+- Fix PlayerController implementation of IAmTargetable.Detectability throwing if no movement component is attached
+- Fix ActorAttackComponent throwing an exception if projectile could not be instantiated (now logs error instead)
+- Add (Try)ForEachEntity extension methods to WorldUtils